OMAP: Fix for bus width which improves SD card's peformance.
This patch improves low speeds for SD cards.
OMAP-MMC controller's can support maximum bus width of '8'.
when bus width is mentioned as "8" in controller data,the SD
stack will check whether bus width is "4" and if not it will
set bus width to "1" and there by degrading performance.
This patch fixes the issue and improves the performance of
SD cards.

Signed-off-by: Kishore Kadiyala <kishore.kadiyala <at> ti.com>
Signed-off-by: Venkatraman S <svenkatr <at> ti.com>
Signed-off-by: Nishanth Menon <nm <at> ti.com>
Acked-by: Madhusudhan Chikkature <madhu.cr <at> ti.com>
Tested-by: Jarkko Nikula <jhnikula <at> gmail.com>
---
In V5 : Rebasing on 2.6.35-rc3
In V4 : Updated with Nishanth's comments and appened his Signed-off
	http://marc.info/?t=127169149300006&r=1&w=2
In V3 : Updated  with Madhu's comments  and appended Tested by Nikula
In V2 : Appended Signed-off by Venkat and Ack by Madhu

Here are my experiment numbers, on a Class 6 SDHC card:
Read peformance is increased by 220%
Write Performance is increased by 52%